+++ extensions/cast/src/main/java/com/google/android/exoplayer2/ext/cast/CastPlayer.java
@@ -276,7 +276,17 @@ public final class CastPlayer extends BasePlayer {
     addListener(eventListener);
   }
 
-  @Override
+  /**
+   * Registers a listener to receive events from the player.
+   *
+   * <p>The listener's methods will be called on the thread associated with {@link
+   * #getApplicationLooper()}.
+   *
+   * @param listener The listener to register.
+   * @deprecated Use {@link #addListener(Listener)} and {@link #removeListener(Listener)} instead.
+   */
+  @Deprecated
+  @SuppressWarnings("deprecation")
   public void addListener(EventListener listener) {
     listeners.add(listener);
   }
@@ -287,7 +297,15 @@ public final class CastPlayer extends BasePlayer {
     removeListener(eventListener);
   }
 
-  @Override
+  /**
+   * Unregister a listener registered through {@link #addListener(EventListener)}. The listener will
+   * no longer receive events from the player.
+   *
+   * @param listener The listener to unregister.
+   * @deprecated Use {@link #addListener(Listener)} and {@link #removeListener(Listener)} instead.
+   */
+  @Deprecated
+  @SuppressWarnings("deprecation")
   public void removeListener(EventListener listener) {
     listeners.remove(listener);
   }